APIO-320 Separate BlogPosting ActionRouter impl and api
